BackEnd/requirements.txt
flask
python-jwt
pymongo
python-dotenv
flask-cors
bson